data-processor/requirements.txt
kafka-python
psycopg2-binary
pandas
pydantic